+++ upload/admin/language/en-gb/common/column_left.php
@@ -78,7 +78,6 @@ $_['text_tax'] = 'Taxes';
 $_['text_tax_class'] = 'Tax Classes';
 $_['text_tax_rate'] = 'Tax Rates';
 $_['text_theme'] = 'Theme Editor';
-$_['text_translation'] = 'Translations';
 $_['text_upgrade'] = 'Upgrade';
 $_['text_upload'] = 'Uploads';
 $_['text_users'] = 'Users';